+++ puma/plot_base.py
@@ -283,7 +283,9 @@ class PlotBase(PlotObject):  # pylint: disable=too-many-instance-attributes
                     "panels are created.",
                     self.n_ratio_panels,
                 )
-            self.fig = Figure(figsize=(8, 6) if self.figsize is None else self.figsize)
+            self.fig = Figure(
+                figsize=(6, 4.5) if self.figsize is None else self.figsize
+            )
             g_spec = gridspec.GridSpec(1, 11, figure=self.fig)
             self.axis_top = self.fig.add_subplot(g_spec[0, :9])
             self.axis_leg = self.fig.add_subplot(g_spec[0, 9:])
@@ -292,14 +294,14 @@ class PlotBase(PlotObject):  # pylint: disable=too-many-instance-attributes
             if self.n_ratio_panels == 0:
                 # no ratio panel
                 self.fig = Figure(
-                    figsize=(8, 6) if self.figsize is None else self.figsize
+                    figsize=(5, 3.5) if self.figsize is None else self.figsize
                 )
                 self.axis_top = self.fig.gca()
 
             elif self.n_ratio_panels == 1:
                 # 1 ratio panel
                 self.fig = Figure(
-                    figsize=(9.352, 6.616) if self.figsize is None else self.figsize
+                    figsize=(5, 4) if self.figsize is None else self.figsize
                 )
 
                 g_spec = gridspec.GridSpec(8, 1, figure=self.fig)