+++ changelog.md
@@ -1,6 +1,6 @@
 # Changelog
 
-### [Latest]
+### [v0.1.5]
 
 - Add `linewidth` and `alpha` to legend handles + set `alpha=1` by default (in `puma.Histogram`) [#92](https://github.com/umami-hep/puma/pull/92)
 - Decreased default `figsize` for plots with zero or one ratio panels [#90](https://github.com/umami-hep/puma/pull/90)